extracted Pick logic and mutation to a separate component from regular Drafting
