+++ client/src/SessionView/Overview/Drafting/Drafting/Drafting.js
@@ -1,23 +1,14 @@
 import { useState, useCallback, useMemo } from 'react'
-import { Grid, Box, Button, Typography } from '@material-ui/core'
+import { Box, Button, Typography } from '@material-ui/core'
 import {
   Block as BanIcon,
   PanTool as PickIcon,
   AssignmentInd as SpeakerIcon,
 } from '@material-ui/icons'
-import Alert from '@material-ui/lab/Alert'
 import shuffle from 'lodash.shuffle'
 import { makeStyles } from '@material-ui/core/styles'
-import clsx from 'clsx'
 
-import speakerFront from '../../../../assets/speaker-front.png'
-import speakerBack from '../../../../assets/speaker-back.png'
 import { useDomainErrors } from '../../../../shared/errorHandling'
-import {
-  ColorBox,
-  getMapPositionName,
-  getMapPositionColor,
-} from '../../../../shared'
 import { MapPreview } from '../../../components'
 import { EditPrompt } from '../../../Edit'
 import { SessionNutshell } from '../../SessionNutshell'
@@ -28,8 +19,9 @@ import { SpeakerIndicator } from '../SpeakerIndicator'
 import { PHASE } from '../shared'
 
 import { BanStepper } from './BanStepper'
-import { PickStepper } from './PickStepper'
 import { ActionOnFactionListButton } from '../components/ActionOnFactionListButton'
+import { Pick } from './Pick'
+import { TablePositionPick } from './TablePositionPick'
 
 function Speaker({ disabled, draft, session, sessionService }) {
   const { setError } = useDomainErrors()
@@ -103,7 +95,7 @@ function Speaker({ disabled, draft, session, sessionService }) {
   )
 }
 
-const useStyles = makeStyles((theme) => ({
+export const useStyles = makeStyles((theme) => ({
   containedButton: {
     transition: theme.transitions.create(
       ['opacity', 'background-color', 'color'],
@@ -138,231 +130,6 @@ const useStyles = makeStyles((theme) => ({
     marginTop: theme.spacing(1),
   },
 }))
-
-function TablePositionPick({
-  disabled,
-  pick,
-  draft,
-  session,
-  selectedPosition,
-  handleSelectedPosition,
-}) {
-  const classes = useStyles()
-
-  return (
-    <Grid
-      className={classes.marginTop}
-      container
-      justifyContent="center"
-      spacing={4}
-    >
-      {[...Array(draft.players.length).keys()].map((tablePositionIndex) => {
-        const picked = draft.picks
-          .filter(({ type }) => type === 'tablePosition')
-          .find(({ pick: p }) => Number(p) === tablePositionIndex)
-        const disabledDueToSelection =
-          selectedPosition !== null && selectedPosition !== tablePositionIndex
-        const isSelected =
-          selectedPosition !== null && selectedPosition === tablePositionIndex
-
-        return (
-          <Grid key={tablePositionIndex} item lg={3} md={4} sm={6} xs={12}>
-            <EditPrompt fullWidth>
-              <Button
-                className={clsx(classes.containedButton, {
-                  [classes.picked]: picked,
-                })}
-                color={isSelected ? 'secondary' : 'default'}
-                disabled={Boolean(
-                  disabled || disabledDueToSelection || picked || pick,
-                )}
-                endIcon={picked ? <PickIcon fontSize="large" /> : null}
-                fullWidth
-                onClick={() => handleSelectedPosition(tablePositionIndex)}
-                variant="contained"
-              >
-                <Typography
-                  style={{
-                    display: 'flex',
-                    alignItems: 'baseline',
-                    gridColumnGap: '0.3em',
-                  }}
-                >
-                  {getMapPositionName({
-                    mapPositions: session.mapPositions,
-                    position: tablePositionIndex,
-                  })}
-                  <ColorBox
-                    color={getMapPositionColor({
-                      mapPositions: session.mapPositions,
-                      position: tablePositionIndex,
-                    })}
-                    disabled
-                  />
-                </Typography>
-                {picked && (
-                  <Typography style={{ marginLeft: '0.3em' }} variant="caption">
-                    picked by {picked.playerName}
-                  </Typography>
-                )}
-              </Button>
-            </EditPrompt>
-          </Grid>
-        )
-      })}
-    </Grid>
-  )
-}
-
-function SpeakerSelectorToggle({ selected, onChange, disabled, cannotSelect }) {
-  const classes = useStyles()
-
-  const onClick = () => {
-    onChange(!selected)
-  }
-
-  return (
-    <Grid
-      className={classes.marginTop}
-      container
-      justifyContent="center"
-      spacing={4}
-    >
-      <Grid item>
-        <div
-          onClick={onClick}
-          style={{
-            cursor: disabled || cannotSelect ? 'auto' : 'pointer',
-            borderRadius: '2%',
-            width: '200px',
-            height: '81px',
-            backgroundSize: '100% auto',
-            border: `2px solid ${selected ? '#f50057' : 'transparent'}`,
-            opacity: !selected || disabled || cannotSelect ? 0.7 : 1,
-            backgroundImage: cannotSelect
-              ? `url(${speakerBack})`
-              : `url(${speakerFront})`,
-          }}
-          title={cannotSelect ? 'already selected' : 'select speaker'}
-        />
-      </Grid>
-    </Grid>
-  )
-}
-
-// TODO onPositionSelected ugly hack because of state being in the wrong place
-// same with onSpeakerSelected (heh)
-function Pick({
-  disabled,
-  pick,
-  clearSelection,
-  draft,
-  session,
-  sessionService,
-  onPositionSelected,
-  onSpeakerSelected,
-}) {
-  const [speakerSelected, setSpeakerSelected] = useState(false)
-  const [selectedPosition, setSelectedPosition] = useState(null)
-  const { setError } = useDomainErrors()
-  const handleSpeakerSelected = useCallback(
-    (v) => {
-      setSpeakerSelected(v)
-      onSpeakerSelected(v)
-    },
-    [onSpeakerSelected],
-  )
-  const handleSelectedPosition = useCallback(
-    (playerIndex) => {
-      const value = selectedPosition === playerIndex ? null : playerIndex
-
-      setSelectedPosition(value)
-      onPositionSelected(value)
-    },
-    [selectedPosition, onPositionSelected],
-  )
-  const pickMutation = async () => {
-    let eventPick = pick || selectedPosition
-    if (speakerSelected) {
-      eventPick = 'speaker'
-    }
-
-    let eventType = pick ? 'faction' : 'tablePosition'
-    if (speakerSelected) {
-      eventType = 'speaker'
-    }
-
-    try {
-      await sessionService.pushEvent(session.id, {
-        type: 'Picked',
-        payload: {
-          sessionId: session.id,
-          pick: eventPick,
-          type: eventType,
-          playerIndex: draft.order[draft.activePlayerIndex],
-          playerName: draft.players[draft.order[draft.activePlayerIndex]],
-        },
-      })
-      setSelectedPosition(null)
-      onPositionSelected(null)
-      handleSpeakerSelected(false)
-      clearSelection()
-    } catch (e) {
-      setError(e)
-    }
-  }
-
-  const { mutate: pickFaction } = useDraftMutation({
-    sessionId: session.id,
-    mutation: pickMutation,
-  })
-
-  return (
-    <>
-      <EditPrompt>
-        <Button
-          color="secondary"
-          disabled={
-            disabled || (!pick && selectedPosition === null && !speakerSelected)
-          }
-          onClick={pickFaction}
-          variant="contained"
-        >
-          pick{' '}
-          {session.setup.options.tablePick ? ' (faction or table spot)' : ''}
-          {session.setup.options.speakerPick ? ' (or speaker)' : ''}
-        </Button>
-      </EditPrompt>
-      {session.setup.options.speakerPick && (
-        <SpeakerSelectorToggle
-          cannotSelect={draft.picks.some(({ type }) => type === 'speaker')}
-          disabled={disabled}
-          onChange={handleSpeakerSelected}
-          selected={speakerSelected}
-        />
-      )}
-      {session.setup.options.tablePick && (
-        <TablePositionPick
-          disabled={
-            disabled ||
-            speakerSelected ||
-            draft.picks.some(
-              ({ type, playerIndex }) =>
-                type === 'tablePosition' &&
-                Number(draft.order[draft.activePlayerIndex]) ===
-                  Number(playerIndex),
-            )
-          }
-          draft={draft}
-          handleSelectedPosition={handleSelectedPosition}
-          pick={pick}
-          selectedPosition={selectedPosition}
-          session={session}
-        />
-      )}
-    </>
-  )
-}
 
 function Ban({
   disabled,
@@ -410,7 +177,6 @@ function Ban({
 }
 
 export function Drafting({ editable, session, sessionService }) {
-  const [disableFactionSelection, setDisableFactionSelection] = useState(false)
   const [selected, setSelected] = useState([])
   const { draft } = useDraftQuery({
     sessionId: session.id,
@@ -481,24 +247,15 @@ export function Drafting({ editable, session, sessionService }) {
         />
       )}
       {draft.phase === PHASE.picks && (
-        <PickStepper draft={draft} mapPositions={session.mapPositions} />
-      )}
-      {draft.phase === PHASE.speaker && (
-        <SpeakerIndicator indicated={draft.speaker} players={draft.players} />
-      )}
-      {draft.phase === PHASE.picks && (
         <Pick
-          clearSelection={() => setSelected([])}
           disabled={!editable}
           draft={draft}
-          onPositionSelected={(selectedPosition) => {
-            setDisableFactionSelection(selectedPosition !== null)
-          }}
-          onSpeakerSelected={setDisableFactionSelection}
-          pick={selected[0]}
           session={session}
           sessionService={sessionService}
         />
+      )}
+      {draft.phase === PHASE.speaker && (
+        <SpeakerIndicator indicated={draft.speaker} players={draft.players} />
       )}
       {draft.phase === PHASE.speaker && (
         <Speaker
@@ -517,7 +274,6 @@ export function Drafting({ editable, session, sessionService }) {
           bans={draft.bans}
           disabled={
             !editable ||
-            disableFactionSelection ||
             draft.picks.some(
               ({ type, playerIndex }) =>
                 type === 'faction' &&
@@ -529,28 +285,6 @@ export function Drafting({ editable, session, sessionService }) {
           max={draft.phase === PHASE.bans ? draft.bansPerRound : 1}
           onSelected={setSelected}
           picks={[]}
-          selected={selected}
-        />
-      )}
-      {draft.phase === PHASE.picks && (
-        <DraftPool
-          bans={[]}
-          disabled={
-            !editable ||
-            disableFactionSelection ||
-            draft.picks.some(
-              ({ type, playerIndex }) =>
-                type === 'faction' &&
-                Number(draft.order[draft.activePlayerIndex]) ===
-                  Number(playerIndex),
-            )
-          }
-          initialPool={draft.initialPool.filter(
-            (factionKey) => !bannedFactionKeys.includes(factionKey),
-          )}
-          max={1}
-          onSelected={setSelected}
-          picks={draft.picks}
           selected={selected}
         />
       )}
